requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.3
python-dotenv==1.0.1
